+++ src/main/java/com/elmakers/mine/bukkit/utility/CompatibilityUtils.java
@@ -50,32 +50,7 @@ public class CompatibilityUtils extends NMSUtils {
      */
     @Deprecated
     public static Location setDirection(Location location, Vector vector) {
-        /*
-         * Sin = Opp / Hyp
-         * Cos = Adj / Hyp
-         * Tan = Opp / Adj
-         *
-         * x = -Opp
-         * z = Adj
-         */
-        final double _2PI = 2 * Math.PI;
-        final double x = vector.getX();
-        final double z = vector.getZ();
-
-        if (x == 0 && z == 0) {
-            location.setPitch(vector.getY() > 0 ? -90 : 90);
-            return location;
-        }
-
-        double theta = Math.atan2(-x, z);
-        location.setYaw((float) Math.toDegrees((theta + _2PI) % _2PI));
-
-        double x2 = NumberConversions.square(x);
-        double z2 = NumberConversions.square(z);
-        double xz = Math.sqrt(x2 + z2);
-        location.setPitch((float) Math.toDegrees(Math.atan(-vector.getY() / xz)));
-
-        return location;
+        return location.setDirection(vector);
     }
 
     public static void applyPotionEffects(LivingEntity entity, Collection<PotionEffect> effects) {